Projet_SSL/im06_ssl_code_grp_3/implémentation_simclr_barlow_twins/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torcheval>=0.0.7
kornia>=0.7.0
opencv-python>=4.8.0
medmnist>=2.2.3
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0
wandb>=0.20.1

# Memory profiling (pour optimisation)
memory-profiler>=0.60.0
psutil>=5.9.0